requirements.txt
streamlit_extras
whisper
audio_recorder_streamlit
azure-cognitiveservices-speech
openai-whisper